[blobfs] (4/4) Convert blobfs to extent-based allocation scheme
This patch allows blobs to be arbitrarily fragmented within blobfs
without API-visible impact on usage. The "TestFragmentation" test
in the blobfs integration test suite is now passing.

ZX-2997 #done
ZX-2968 #comment In Progress
PKG-313 #comment In Progress

Test: /boot/test/fs/blobfs-unit-tests,
      /boot/test/fs/blobfs-integration-tests

+++ system/ulib/blobfs/include/blobfs/common.h
@@ -43,7 +43,7 @@ zx_status_t CheckSuperblock(const Superblock* info, uint64_t max);
 zx_status_t GetBlockCount(int fd, uint64_t* out);
 int Mkfs(int fd, uint64_t block_count);
 
-uint64_t MerkleTreeBlocks(const Inode& blobNode);
+uint32_t MerkleTreeBlocks(const Inode& blobNode);
 
 // Get a pointer to the nth block of the bitmap.
 inline void* GetRawBitmapData(const RawBitmap& bm, uint64_t n) {
